cloudinary file uplode correction

+++ src/utils/cloudinary.js
@@ -1,5 +1,4 @@
 import {v2 as cloudinary} from "cloudinary"
-import { log } from "console";
 import fs from "fs"
 
 //configuration of cloudnery
@@ -12,7 +11,7 @@ cloudinary.config({
     });
 
 
-const uploadFileToCloudinary= async (localFilePath)=>{
+const uploadOnCloudinary= async (localFilePath)=>{
     try {
         if(!localFilePath)return null
         //upload file in cloudinary
@@ -21,35 +20,38 @@ const uploadFileToCloudinary= async (localFilePath)=>{
         })
         //file has been uploaded successfully
         // console.log('file is uploaded at cloudinary',response.url);
+        console.log("hey i am uploading....")
 
-         fs.unlinkSync(localFilePath)
+        //  fs.unlinkSync(localFilePath)
 
         //code from the chatgpt
 
          // delete local file after success
-    // if (fs.existsSync(localFilePath)) {
-    //   fs.unlinkSync(localFilePath);
-    // }
+    if (fs.existsSync(localFilePath)) {
+      fs.unlinkSync(localFilePath);
+    }
 
         return response
         
 
         
     } catch (error) {
+
+        console.log("hey i am not uploading....")
         // // if the file has issue in uploading into the server otherwise file is not uploading then unlink the file into the local storage 
-        fs.unlinkSync(localFilePath)
+        // fs.unlinkSync(localFilePath)
         // //this command is remove the file which is temporary store into the localstorage 
 
         //code from the chatgpt
 
     //    // delete local file if exists
-    // if (fs.existsSync(localFilePath)) {
-    //   fs.unlinkSync(localFilePath);
-    // }
+    if (fs.existsSync(localFilePath)) {
+      fs.unlinkSync(localFilePath);
+    }
 
     return null;
     }
 
 }
 
-export {uploadFileToCloudinary}
+export {uploadOnCloudinary}
